+++ app/src/androidTest/java/paperdb/io/paperdb/ReadBenchmark.java
@@ -1,7 +1,7 @@
 package paperdb.io.paperdb;
 
 import android.os.SystemClock;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import android.util.Log;
 
 import org.junit.Before;
@@ -17,7 +17,7 @@ import io.paperdb.Paper;
 import paperdb.io.paperdb.testdata.Person;
 import paperdb.io.paperdb.testdata.TestDataGenerator;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.InstrumentationRegistry.getTargetContext;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(AndroidJUnit4.class)