+++ paperdb/src/androidTest/java/io/paperdb/CompatibilityTest.java
@@ -1,6 +1,6 @@
 package io.paperdb;
 
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -8,7 +8,7 @@ import org.junit.runner.RunWith;
 
 import java.lang.reflect.Field;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.InstrumentationRegistry.getTargetContext;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**