+++ paperdb/src/androidTest/java/io/paperdb/deprecated/PaperDeprecatedAPITest.java
@@ -1,6 +1,6 @@
 package io.paperdb.deprecated;
 
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -12,7 +12,7 @@ import io.paperdb.Paper;
 import io.paperdb.PaperDbException;
 import io.paperdb.testdata.TestDataGenerator;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.InstrumentationRegistry.getTargetContext;
 import static junit.framework.TestCase.assertTrue;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertFalse;